requirements.txt
fastapi
sqlmodel
uvicorn
python-dotenv
passlib[bcrypt]
python-jose
psycopg2-binary
